Calculation of user points

+++ src/main/java/com/majorkeytech/hackaton/model/User.java
@@ -23,5 +23,5 @@ public class User {
     private String lastName;
     private String email;
     private String password;
-    private Long points; 
+    private int points; 
 }
